requirements.txt
Django==3.2.6
Pillow==8.3.1
python-dotenv==0.19.0
social-auth-app-django==5.0.0
django-countries